consolidate routes and components from audrey and seans
